backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
aiomysql>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain-core>=0.1.9
langchain-community>=0.0.12
langchain-google-genai>=0.0.5